[FIX] Nest and shooting.

+++ AIFinal/CollisionDetection.h
@@ -6,7 +6,7 @@ class CollisionDetection
 public:
 	CollisionDetection();
 	~CollisionDetection();
-	const float PI = 3.14159265359;
+	const double PI = 3.14159265359;
 
 	void CircleTest(GameObject object1, GameObject object2);
 	void BoxTest(GameObject object1, GameObject object2);
